Enum bitreader::BitReaderError

pub enum BitReaderError {
    NotEnoughData {
        position: u64,
        length: u64,
        requested: u64,
    },
    TooManyBitsForType {
        position: u64,
        requested: u8,
        allowed: u8,
    },
}

Error enumeration of BitReader errors.

Variants
NotEnoughData

Requested more bits than there are left in the byte slice at the current position.

Fields

position: u64

length: u64

requested: u64

TooManyBitsForType

Requested more bits than the returned variable can hold, for example more than 8 bits when reading into a u8.

Fields

position: u64

requested: u8

allowed: u8